+++ tests/testthat/test-bioc.R
@@ -36,7 +36,7 @@ test_that("internal map is current", {
   skip_if_offline()
   expect_equal(
     bioconductor$get_release_version(),
-    package_version("3.20"))
+    package_version("3.23"))
 })
 
 test_that("set of repos are correct", {