+++ tests/testthat/test-system.R
@@ -1,12 +1,20 @@
 test_that("system_check", {
 
-  mockery::stub(system_check, "system2", structure("output", status = 1))
+  local_mocked_bindings(
+    system2 = function(...) structure("output", status = 1),
+    .package = "base"
+  )
+
   expect_error(
     system_check("foobar", args = c("arg1", "arg2", quiet = TRUE)),
     "Command foobar failed"
   )
 
-  mockery::stub(system_check, "system2", 42)
+  local_mocked_bindings(
+    system2 = function(...) 42,
+    .package = "base"
+  )
+
   expect_error(
     system_check("foobar", args = c("arg1", "arg2", quiet = FALSE)),
     "Command foobar failed"